Refactors
Refactored Applayout to SyllabusLayout, this is becaue there is no defined app layout and only a layout for the syllabus
