+++ frontend/src/screens/SyllabusView/CourseSchedule.tsx
@@ -1,6 +1,6 @@
 import {useNavigate, useLocation} from "react-router-dom";
 import {handleBack,handleNext} from "../../components/Button/ButtonLogic";
-import AppLayout from "../../ApplicationLayout/Applayout";
+import AppLayout from "../../SyllabusLayout/SyllabusLayout";
 import {useState} from "react";
 
 const CourseSchedule = () => {
